{invar,blacksteel}: lock dbus impl to dbus-broker
This should void future disturbance of switching default impl. A recent
accident[1] shows that reversion PR is not treated serious enough  as
another big change that need extensive testing and compatibility check
(even when being warned[2] of potential issues, which is exactly what
happens later). It works previously does not imply *changing* it, either
forward or back, works as well.

So here we lock it from changes without human intervention. The exact
implementation does not matter to us. The newer dbus-broker is used
because why not.

[1]: https://github.com/NixOS/nixpkgs/pull/307731
[2]: https://github.com/NixOS/nixpkgs/pull/307731#issuecomment-2082442595